+++ Paylocity/Handlers/Interfaces/IRulesHandler.cs
@@ -4,6 +4,6 @@ namespace Paylocity.Handlers.Interfaces
     public interface IRulesHandler
     {
         bool IsDiscounted(string name);
-        decimal GetCostAnnual(string name, bool isDependent);
+        double GetCostAnnual(string name, bool isDependent);
     }
 }
